requirements.txt
librosa
matplotlib
numpy
pandas
scikit-learn
tensorflow-gpu
tqdm